+++ backend/core/src/listings/listings.controller.ts
@@ -8,14 +8,28 @@ import {
   Post,
   Put,
   Query,
+  Req,
   UseGuards,
   UseInterceptors,
   UsePipes,
   ValidationPipe,
 } from "@nestjs/common"
+import { Request } from "express"
 import { ListingsService } from "./listings.service"
-import { ApiBearerAuth, ApiOperation, ApiTags } from "@nestjs/swagger"
-import { ListingCreateDto, ListingDto, ListingUpdateDto } from "./dto/listing.dto"
+import {
+  ApiBearerAuth,
+  ApiExtraModels,
+  ApiOperation,
+  ApiQuery,
+  ApiTags,
+  getSchemaPath,
+} from "@nestjs/swagger"
+import {
+  ListingCreateDto,
+  ListingDto,
+  ListingUpdateDto,
+  ListingFilterParams,
+} from "./dto/listing.dto"
 import { ResourceType } from "../auth/decorators/resource-type.decorator"
 import { OptionalAuthGuard } from "../auth/guards/optional-auth.guard"
 import { AuthzGuard } from "../auth/guards/authz.guard"
@@ -39,9 +53,30 @@ export class ListingsController {
     required: false,
     type: String,
   })
+  @ApiExtraModels(ListingFilterParams)
+  @ApiQuery({
+    name: "filter",
+    required: false,
+    type: [String],
+    schema: {
+      type: "array",
+      example: [
+        { $comparison: "=", status: "active" },
+        { $comparison: "<>", name: "Coliseum" },
+      ],
+      items: {
+        $ref: getSchemaPath(ListingFilterParams),
+      },
+    },
+  })
   @UseInterceptors(CacheInterceptor)
-  public async getAll(@Query("jsonpath") jsonpath?: string): Promise<ListingDto[]> {
-    return mapTo(ListingDto, await this.listingsService.list(jsonpath))
+  public async getAll(
+    @Req() request: Request,
+    @Query("jsonpath") jsonpath?: string,
+    @Query("filter") filter?: ListingFilterParams[]
+    // TODO: Add options param here for paging and sorting
+  ): Promise<ListingDto[]> {
+    return mapTo(ListingDto, await this.listingsService.list(jsonpath, filter))
   }
 
   @Post()